texture_cache: Split templates out
